+++ DTCD-ExtensionRiskPrimitives/src/primitives/RiskNode/RiskNode.js
@@ -14,6 +14,9 @@ export default class RiskNode extends NodeWithDefaultLabel {
   constructor(yFiles) {
     super(yFiles)
     this.yfiles = yFiles.default;
+
+    this.instance.tag.defaultLabel = `$this.props.name$`
+    this.instance.tag.defaulInitialtLabel = ``
   }
 
   create() {