+++ CHANGELOG.md
@@ -26,7 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - new RiskRAppetiteNode primitive
 - default label for nodes
 - AbstractNode class
-- NodeWithDefaultLabel class
+- NodeWithDefaultLabel class 
 
 ## [0.2.0]